Move track filtering widgets to a separate .h/.cpp
Introduce files "trackfilterwidgets.h" and "trackfilterwidgets.cpp" and
move the following classes there:

 - FilterPickerWidget
 - FilterLineWidget
 - FiltersListWidget
